Add `CorePlayer` and modify `VideoPlayer` as a ready to go player
