+++ src/components/bottom-control-buttons/components/PlayPauseReplay.tsx
