+++ src/components/bottom-control-buttons/components/VolumeBarStyled.ts
