+++ src/components/bottom-control-buttons/components/VolumeSlider.tsx
