+++ src/components/bottom-control-buttons/hooks/useTimeDisplayStyles.ts
