+++ src/components/controls/Controls.tsx
@@ -1,28 +1,14 @@
-import { FC } from 'react';
-
-import { BottomControls } from '../bottom-controls/BottomControls';
-import { CenteredBottomPlayback } from '../centered-bottom-playback/CenteredBottomPlayback';
-import { CenteredPlayButton } from '../centered-play-button/CenteredPlayButton';
-import { CenteredReplayButton } from '../centered-replay-button/CenteredReplayButton';
-import { PauseAnimation } from '../play-pause-animation/PauseAnimation';
-import { PlayAnimation } from '../play-pause-animation/PlayAnimation';
+import { FC, ReactNode } from 'react';
 
 import { useControlsStyles } from './useControlsStyles';
 
-export interface ControlProps {}
+export interface ControlProps {
+	children: ReactNode;
+}
 
-export const Controls: FC<ControlProps> = () => {
+export const Controls: FC<ControlProps> = ({ children }) => {
 	// Controls styles
-	const { wrapper } = useControlsStyles().classes;
+	const { controls } = useControlsStyles().classes;
 
-	return (
-		<div className={wrapper}>
-			<PlayAnimation />
-			<PauseAnimation />
-			<CenteredPlayButton />
-			<CenteredBottomPlayback />
-			<CenteredReplayButton />
-			<BottomControls />
-		</div>
-	);
+	return <div className={controls}>{children}</div>;
 };
