+++ src/components/controls/useControlsStyles.ts
@@ -1,7 +1,7 @@
 import { makeStyles } from 'tss-react/mui';
 
 export const useControlsStyles = makeStyles()(() => ({
-	wrapper: {
+	controls: {
 		width: '100%',
 		height: '100%',
 		position: 'absolute',
