+++ src/components/index.ts
@@ -1,2 +1,22 @@
 export * from './video-player/VideoPlayer';
-export * from './video-player/VideoContainer';
+export * from './animated-icon-wrapper/AnimatedIconWrapper';
+export * from './big-centered-button/BigCenteredButton';
+export * from './bottom-control-buttons/BottomControlButtons';
+export * from './bottom-control-buttons/components';
+export * from './bottom-controls/BottomControls';
+export * from './centered-bottom-playback/CenteredBottomPlayback';
+export * from './centered-bottom-playback/components/PlaybackRateButtonStyled';
+export * from './centered-play-button/CenteredPlayButton';
+export * from './centered-replay-button/CenteredReplayButton';
+export * from './controls/Controls';
+export * from './core-player/CorePlayer';
+export * from './core-player/types';
+export * from './draggable-popover/DraggablePopover';
+export * from './icons';
+export * from './pip-controls/PipControls';
+export * from './play-pause-animation/PauseAnimation';
+export * from './play-pause-animation/PlayAnimation';
+export * from './progress-bar/ProgressBar';
+export * from './video-container/VideoContainer';
+export * from './video-player/VideoPlayer';
+export * from './video-poster/VideoPoster';
