+++ src/components/video-player/useVideoPlayerStyles.ts
@@ -1,0 +1,9 @@
+import { makeStyles } from 'tss-react/mui';
+
+export const useVideoPlayerStyles = makeStyles()({
+	gridCentered: {
+		display: 'inline-flex',
+		height: '100%',
+		alignItems: 'center',
+	},
+});
