+++ src/types/video-state.ts
@@ -1,12 +1,8 @@
 import { Emitter } from 'mitt';
-import { Dispatch, MutableRefObject, ReactNode, RefObject } from 'react';
+import { Dispatch, MutableRefObject, RefObject } from 'react';
 import type ReactPlayer from 'react-player';
 
-import { VideoContext } from '../context';
-import { BlendColors } from '../utils/colors';
-
 import {
-	ControlsConfig,
 	EmitterAddRemoveListeners,
 	EmitterEvents,
 	VideoActions,
@@ -39,41 +35,6 @@ export interface Highlight extends Segment {
 	id: string;
 	/** Color of the highlight. This must be a HEX color code */
 	colors: string[];
-}
-
-/** Provider's initialization state */
-interface VideoPlayerInitialState {
-	/** If the video start playing from start */
-	playing?: boolean;
-	/** Time (in ms) that video will start to play */
-	startTime: number;
-	/** Time (in ms) that video should pause */
-	endTime: number;
-	/** Video duration */
-	duration: number;
-	/** Current played time */
-	currentTime: number;
-	highlights: Highlight[];
-}
-
-/**
- * Context Provider for playing videos
- */
-
-export interface VideoProviderProps {
-	/** Configuration that enables/disables some parts of the overlay on top of the video player */
-	controlsConfig?: ControlsConfig;
-	/** Provider's initialization state */
-	initialState?: VideoPlayerInitialState;
-	/** ReactNode that will consume the context */
-	children: ReactNode;
-	/** State that needs to be stored in localStorage */
-	persistedState?: VideoState;
-	/** Blending colors for highlights presented in `<ProgressBar>` */
-	getHighlightColorBlended?: BlendColors;
-	/** A callback that can updates VideoContext outside of the VideoProvider */
-	onContext?: (context: VideoContext) => void;
-	highlights?: Highlight[];
 }
 
 /**
