+++ src/utils/testing-render.tsx
@@ -2,9 +2,12 @@ import { ThemeProvider } from '@mui/material/styles';
 import { render as renderRTL, RenderOptions } from '@testing-library/react';
 import { FC, ReactNode, useRef } from 'react';
 
-import { DEFAULT_CONTROLS_CONFIG } from '../components/controls/controls-config';
-import { VideoContext, VideoProvider } from '../context';
+import { PROVIDER_INITIAL_STATE } from '../components/core-player/types';
+import { VideoContext } from '../context/video';
+import { VideoProvider } from '../context/VideoProvider';
 import { createPlayerTheme } from '../theme';
+
+import { blend } from './colors';
 export * from '@testing-library/react';
 export { default as userEvent } from '@testing-library/user-event';
 
@@ -16,7 +19,10 @@ export const renderWithProviders = (
 	options?: RenderOptions,
 ) =>
 	renderRTL(
-		<VideoProvider controlsConfig={DEFAULT_CONTROLS_CONFIG}>
+		<VideoProvider
+			initialState={PROVIDER_INITIAL_STATE}
+			getHighlightColorBlended={blend}
+		>
 			<ThemeProvider theme={playerTheme}>{Component}</ThemeProvider>
 		</VideoProvider>,
 		options,
