--- src/utils/throttle.ts
@@ -1,11 +1,0 @@
-export function throttle(callback: VoidFunction, interval: number) {
-	let enableCall = true;
-
-	return (...args: any[]) => {
-		if (!enableCall) return;
-
-		enableCall = false;
-		callback.apply(args);
-		setTimeout(() => (enableCall = true), interval);
-	};
-}
